Creacion de Model Base para poder extender todas las variables para Auditoria
