+++ src/main/java/com/example/taskmanager/category/Category.java
@@ -1,26 +1,22 @@
 package com.example.taskmanager.category;
 
+import com.example.taskmanager.modelbase.ModelBase;
+
 import java.sql.Timestamp;
 
-public class Category {
+public class Category extends ModelBase {
     private String categoryId;
     private String name;
     private String description;
-    private String uuid;
-    private Timestamp createdDate;
-    private Timestamp modifiedDate;
-    private Integer createdBy;
-    private Integer modifiedBy;
-    private boolean deleted;
 
     public Category(String uuid) {
-        this.uuid=uuid;
+        super(uuid);
     }
 
     public Category(String name, String description, String uuid) {
+        super(uuid);
         this.name = name;
         this.description = description;
-        this.uuid=uuid;
     }
 
     public String getCategoryId() {
@@ -47,51 +43,4 @@ public class Category {
         this.description = description;
     }
 
-    public String getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(String uuid) {
-        this.uuid = uuid;
-    }
-
-    public Timestamp getCreatedDate() {
-        return createdDate;
-    }
-
-    public void setCreatedDate(Timestamp createdDate) {
-        this.createdDate = createdDate;
-    }
-
-    public Timestamp getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Timestamp modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public Integer getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(Integer createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public Integer getModifiedBy() {
-        return modifiedBy;
-    }
-
-    public void setModifiedBy(Integer modifiedBy) {
-        this.modifiedBy = modifiedBy;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
-    }
 }
