+++ src/main/java/com/example/taskmanager/status/Status.java
@@ -1,26 +1,21 @@
 package com.example.taskmanager.status;
 
 
+import com.example.taskmanager.modelbase.ModelBase;
+
 import java.sql.Timestamp;
 
-public class Status {
+public class Status extends ModelBase {
     private String stateId;
     private String name;
 
-    private String uuid;
-    private Timestamp createdDate;
-    private Timestamp modifiedDate;
-    private Integer createdBy;
-    private Integer modifiedBy;
-    private boolean deleted;
-
     public Status(String uuid) {
-        this.uuid = uuid;
+        super(uuid);
     }
 
     public Status(String name, String uuid) {
+        super(uuid);
         this.name = name;
-        this.uuid = uuid;
     }
 
     public String getStateId() {
@@ -39,51 +34,5 @@ public class Status {
         this.name = name;
     }
 
-    public String getUuid() {
-        return uuid;
-    }
 
-    public void setUuid(String uuid) {
-        this.uuid = uuid;
-    }
-
-    public Timestamp getCreatedDate() {
-        return createdDate;
-    }
-
-    public void setCreatedDate(Timestamp createdDate) {
-        this.createdDate = createdDate;
-    }
-
-    public Timestamp getModifiedDate() {
-        return modifiedDate;
-    }
-
-    public void setModifiedDate(Timestamp modifiedDate) {
-        this.modifiedDate = modifiedDate;
-    }
-
-    public Integer getCreatedBy() {
-        return createdBy;
-    }
-
-    public void setCreatedBy(Integer createdBy) {
-        this.createdBy = createdBy;
-    }
-
-    public Integer getModifiedBy() {
-        return modifiedBy;
-    }
-
-    public void setModifiedBy(Integer modifiedBy) {
-        this.modifiedBy = modifiedBy;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
-    }
 }
